+++ src/Firebase/Messaging/SendReport.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Kreait\Firebase\Messaging;
 
+use Kreait\Firebase\Exception\Messaging\InvalidMessage;
+use Kreait\Firebase\Exception\Messaging\NotFound;
 use Throwable;
 
 class SendReport
@@ -52,6 +54,23 @@ class SendReport
     public function isFailure(): bool
     {
         return $this->error !== null;
+    }
+
+    public function messageTargetWasInvalid(): bool
+    {
+        $errorMessage = $this->error ? $this->error->getMessage() : '';
+
+        return $this->messageWasInvalid() && \preg_match('/((not.+valid)|invalid).+token/i', $errorMessage) === 1;
+    }
+
+    public function messageWasInvalid(): bool
+    {
+        return $this->error instanceof InvalidMessage;
+    }
+
+    public function messageWasSentToUnknownToken(): bool
+    {
+        return $this->error instanceof NotFound;
     }
 
     /**